email_service/airflow/requirements.txt
apache-airflow==1.10.12
bcrypt==3.2.0
cattrs
cffi==1.14.3
click==7.1.2
attrs
cryptography==3.1.1
dataclasses==0.6
fastapi==0.65.2
Flask==1.1.2
h11==0.9.0
httptools==0.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
paramiko==2.7.2
psycopg2-binary==2.8.5
pycparser==2.20
pydantic==1.8.2
PyNaCl==1.4.0
six==1.15.0
starlette==0.14.2
tqdm==4.49.0
uvicorn==0.14.0
uvloop==0.14.0
websockets==8.1
Werkzeug==0.16.1
SQLAlchemy==1.3.23
